# server/openapi.yaml
openapi: 3.0.0
info:
  title: Mini Supermarket API
  version: 1.0.0
  description: API cho hệ thống quản lý siêu thị mini
servers:
  - url: 'http://localhost:5000'
...
paths:
...
  '/api/carts/customer/{customerId}':
    get:
      tags:
        - cart
      summary: get Cart By Customer
      operationId: getCartByCustomer
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/carts/:
    get:
      tags:
        - cart
      summary: get All Carts
      operationId: getAllCarts
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{id}':
    get:
      tags:
        - cart
      summary: get Cart By Id
      operationId: getCartById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{cartId}/items':
    post:
      tags:
        - cart
      summary: add Item To Cart
      operationId: addItemToCart
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/items/{itemId}/quantity':
    put:
      tags:
        - cart
      summary: update Item Quantity
      operationId: updateItemQuantity
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/items/{itemId}':
    delete:
      tags:
        - cart
      summary: remove Item From Cart
      operationId: removeItemFromCart
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{cartId}/apply-promo':
    post:
      tags:
        - cart
      summary: apply Promo
      operationId: applyPromo
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{cartId}/remove-promo':
    delete:
      tags:
        - cart
      summary: remove Promo
      operationId: removePromo
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{cartId}/clear':
    delete:
      tags:
        - cart
      summary: clear Cart
      operationId: clearCart
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/carts/stats:
    get:
      tags:
        - cart
      summary: get Cart Stats
      operationId: getCartStats
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/carts/{cartId}/checkout':
    patch:
      tags:
        - cart
      summary: checkout Cart
      operationId: checkoutCart
      parameters:
        - in: path
          name: cartId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
...